Refactoring application initialisatoin
